Refactor http transport into separate file.
